refactors routes file into a folder with files; adds a helper routes file
