+++ app/__init__.py
@@ -4,10 +4,10 @@ from flask_migrate import Migrate
 import os
 from dotenv import load_dotenv
 
+
 db = SQLAlchemy()
 migrate = Migrate()
 load_dotenv()
-
 
 def create_app(test_config=None):
     app = Flask(__name__)
@@ -21,18 +21,15 @@ def create_app(test_config=None):
         app.config["SQLALCHEMY_DATABASE_URI"] = os.environ.get(
             "SQLALCHEMY_TEST_DATABASE_URI")
 
-    # Import models here for Alembic setup
-    from app.models.task import Task
-    from app.models.goal import Goal
+    # from app.models.task import Task
+    # from app.models.goal import Goal
 
     db.init_app(app)
     migrate.init_app(app, db)
 
-    # Register Blueprints here
-    from .routes import tasks_bp
-    from .routes import goals_bp
+    from .routes.task_routes import tasks_bp
+    from .routes.goal_routes import goals_bp
     app.register_blueprint(tasks_bp)
     app.register_blueprint(goals_bp)
-
 
     return app
