+++ app/models/goal.py
@@ -23,8 +23,5 @@ class Goal(db.Model):
         tasks_list = []
         for task in self.tasks:
             tasks_list.append(task.to_dict_with_goal_id())
-        # tasks_id_list = self.get_task_ids()
-        # for ea_id in tasks_id_list:
-        #     tasks_list.append(Task.query.get(ea_id))
         return tasks_list
